requirements.txt
anyio==3.6.2
certifi==2023.5.7
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
loguru==0.7.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
python-telegram-bot==20.3
sniffio==1.3.0
